[#52] music repository interface 삭제
