+++ app/src/main/java/com/soi/moya/models/Version.kt
@@ -1,0 +1,7 @@
+package com.soi.moya.models
+
+data class Version(
+    val version: String = "",
+    val isRequired: Boolean = false,
+    val feature: List<String>
+)
